# docs/swagger.yaml
basePath: /
definitions:
...
  domain.Product:
    properties:
      category_id:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      image:
        type: string
      name:
        type: string
      price:
        type: number
      updated_at:
        type: string
    type: object
...
  dto.CreateProductRequest:
    properties:
      category_id:
        type: string
      description:
        type: string
      image:
        type: string
      name:
        type: string
      price:
        type: number
    type: object
...
info:
  contact: {}
  description: APIs for using the management system and sales orders
  title: Skina Lanches Management API
  version: "1.0"
paths:
...
  /products:
    get:
      consumes:
      - application/json
      description: Retrieves a paginated list of products optionally filtered by category.
      parameters:
      - description: Filter products by category
        in: query
        name: category
        type: string
      - default: 1
        description: Page number for pagination
        in: query
        name: page
        type: integer
      - default: 10
        description: Number of products per page
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved list of products
          schema:
            items:
              $ref: '#/definitions/domain.Product'
            type: array
        "500":
          description: Internal server error if there is a problem on the server side
      summary: List products
      tags:
      - products
    post:
      consumes:
      - application/json
      description: Adds a new product to the database with the given details.
      parameters:
      - description: Product creation details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateProductRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Product successfully created
          schema:
            $ref: '#/definitions/domain.Product'
        "400":
          description: Bad request if the product data is invalid
        "500":
          description: Internal server error if there is a problem on the server side
      summary: Add a new product
      tags:
      - products
  /products/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes a product based on its unique ID and returns a success
        message.
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Message indicating successful deletion
        "400":
          description: Bad request if the ID is not provided or is invalid
        "404":
          description: Product not found if the ID does not match any product
        "500":
          description: Internal server error if there is a problem deleting the product
      summary: Delete a product
      tags:
      - products
    get:
      consumes:
      - application/json
      description: Retrieves details of a product based on its unique ID.
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved the product details
          schema:
            $ref: '#/definitions/domain.Product'
        "400":
          description: Bad request if the ID is not provided or invalid
        "404":
          description: Product not found if the ID does not match any product
      summary: Get a product
      tags:
      - products
    patch:
      consumes:
      - application/json
      description: Replace product details in the database by ID.
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      - description: Product object that needs to be replaced
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Product successfully updated
          schema:
            $ref: '#/definitions/domain.Product'
        "400":
          description: Invalid input, Object is invalid
          schema:
            type: string
        "404":
          description: Product not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Replace an existing product
      tags:
      - products
    put:
      consumes:
      - application/json
      description: Update product details in the database by ID.
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      - description: Product object that needs to be updated
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Product successfully updated
          schema:
            $ref: '#/definitions/domain.Product'
        "400":
          description: Invalid input, Object is invalid
          schema:
            type: string
        "404":
          description: Product not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Update an existing product
      tags:
      - products
swagger: "2.0"